separate out LifecyclePolicy#toSteps from Policy
