+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/indexlifecycle/LifecyclePolicyTests.java
@@ -111,86 +111,6 @@ public class LifecyclePolicyTests extends AbstractSerializingTestCase<LifecycleP
         assertSame(TimeseriesLifecycleType.INSTANCE, policy.getType());
     }
 
-    public void testFirstAndLastSteps() {
-        Client client = mock(Client.class);
-        LongSupplier nowSupplier = () -> 0L;
-        lifecycleName = randomAlphaOfLengthBetween(1, 20);
-        Map<String, Phase> phases = new LinkedHashMap<>();
-        LifecyclePolicy policy = new LifecyclePolicy(TestLifecycleType.INSTANCE, lifecycleName, phases);
-        List<Step> steps = policy.toSteps(client, nowSupplier);
-        assertThat(steps.size(), equalTo(2));
-        assertThat(steps.get(0), instanceOf(InitializePolicyContextStep.class));
-        assertThat(steps.get(0).getKey(), equalTo(new StepKey("pre-phase", "pre-action", "init")));
-        assertThat(steps.get(0).getNextStepKey(), equalTo(TerminalPolicyStep.KEY));
-        assertSame(steps.get(1), TerminalPolicyStep.INSTANCE);
-    }
-
-    public void testToStepsWithOneStep() {
-        Client client = mock(Client.class);
-        LongSupplier nowSupplier = () -> 0L;
-        MockStep mockStep = new MockStep(
-            new Step.StepKey("test", "test", "test"), TerminalPolicyStep.KEY);
-
-        lifecycleName = randomAlphaOfLengthBetween(1, 20);
-        Map<String, Phase> phases = new LinkedHashMap<>();
-        LifecycleAction firstAction = new MockAction(Arrays.asList(mockStep));
-        Map<String, LifecycleAction> actions = Collections.singletonMap(MockAction.NAME, firstAction);
-        Phase firstPhase = new Phase("test", TimeValue.ZERO, actions);
-        phases.put(firstPhase.getName(), firstPhase);
-        LifecyclePolicy policy = new LifecyclePolicy(TestLifecycleType.INSTANCE, lifecycleName, phases);
-        StepKey firstStepKey = InitializePolicyContextStep.KEY;
-        StepKey secondStepKey = new StepKey("new", PhaseAfterStep.NAME, PhaseAfterStep.NAME);
-        List<Step> steps = policy.toSteps(client, nowSupplier);
-        assertThat(steps.size(), equalTo(4));
-        assertSame(steps.get(0).getKey(), firstStepKey);
-        assertThat(steps.get(0).getNextStepKey(), equalTo(secondStepKey));
-        assertThat(steps.get(1).getKey(), equalTo(secondStepKey));
-        assertThat(steps.get(1).getNextStepKey(), equalTo(mockStep.getKey()));
-        assertThat(steps.get(2).getKey(), equalTo(mockStep.getKey()));
-        assertThat(steps.get(2).getNextStepKey(), equalTo(TerminalPolicyStep.KEY));
-        assertSame(steps.get(3), TerminalPolicyStep.INSTANCE);
-    }
-
-    public void testToStepsWithTwoPhases() {
-        Client client = mock(Client.class);
-        LongSupplier nowSupplier = () -> 0L;
-        MockStep secondActionStep = new MockStep(new StepKey("second_phase", "test2", "test"), TerminalPolicyStep.KEY);
-        MockStep secondAfter = new MockStep(new StepKey("first_phase", PhaseAfterStep.NAME, PhaseAfterStep.NAME),
-                secondActionStep.getKey());
-        MockStep firstActionAnotherStep = new MockStep(new StepKey("first_phase", "test", "bar"), secondAfter.getKey());
-        MockStep firstActionStep = new MockStep(new StepKey("first_phase", "test", "foo"), firstActionAnotherStep.getKey());
-        MockStep firstAfter = new MockStep(new StepKey("new", PhaseAfterStep.NAME, PhaseAfterStep.NAME), firstActionStep.getKey());
-        MockStep init = new MockStep(InitializePolicyContextStep.KEY, firstAfter.getKey());
-
-        lifecycleName = randomAlphaOfLengthBetween(1, 20);
-        Map<String, Phase> phases = new LinkedHashMap<>();
-        LifecycleAction firstAction = new MockAction(Arrays.asList(firstActionStep, firstActionAnotherStep));
-        LifecycleAction secondAction = new MockAction(Arrays.asList(secondActionStep));
-        Map<String, LifecycleAction> firstActions = Collections.singletonMap(MockAction.NAME, firstAction);
-        Map<String, LifecycleAction> secondActions = Collections.singletonMap(MockAction.NAME, secondAction);
-        Phase firstPhase = new Phase("first_phase", TimeValue.ZERO, firstActions);
-        Phase secondPhase = new Phase("second_phase", TimeValue.ZERO, secondActions);
-        phases.put(firstPhase.getName(), firstPhase);
-        phases.put(secondPhase.getName(), secondPhase);
-        LifecyclePolicy policy = new LifecyclePolicy(TestLifecycleType.INSTANCE, lifecycleName, phases);
-
-        List<Step> steps = policy.toSteps(client, nowSupplier);
-        assertThat(steps.size(), equalTo(7));
-        assertThat(steps.get(0).getClass(), equalTo(InitializePolicyContextStep.class));
-        assertThat(steps.get(0).getKey(), equalTo(init.getKey()));
-        assertThat(steps.get(0).getNextStepKey(), equalTo(init.getNextStepKey()));
-        assertThat(steps.get(1).getClass(), equalTo(PhaseAfterStep.class));
-        assertThat(steps.get(1).getKey(), equalTo(firstAfter.getKey()));
-        assertThat(steps.get(1).getNextStepKey(), equalTo(firstAfter.getNextStepKey()));
-        assertThat(steps.get(2), equalTo(firstActionStep));
-        assertThat(steps.get(3), equalTo(firstActionAnotherStep));
-        assertThat(steps.get(4).getClass(), equalTo(PhaseAfterStep.class));
-        assertThat(steps.get(4).getKey(), equalTo(secondAfter.getKey()));
-        assertThat(steps.get(4).getNextStepKey(), equalTo(secondAfter.getNextStepKey()));
-        assertThat(steps.get(5), equalTo(secondActionStep));
-        assertSame(steps.get(6), TerminalPolicyStep.INSTANCE);
-    }
-
     public void testIsActionSafe() {
         Map<String, Phase> phases = new LinkedHashMap<>();
         LifecycleAction firstAction = new MockAction(Collections.emptyList(), true);
@@ -493,6 +413,11 @@ public class LifecyclePolicyTests extends AbstractSerializingTestCase<LifecycleP
                     }
                 });
             });
+        }
+
+        @Override
+        public Map<String, StepsFactory.ActionStepsCreator<?>> getActionStepsCreators() {
+            return null;
         }
     }
 }
